# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salons/{salonId}/nail-colors:
    get:
      operationId: listNailColors
      tags: [nail-colors]
      summary: List nail color inventory
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: brand
          schema:
            type: string
        - in: query
          name: finish
          schema:
            type: string
            enum: [glossy, matte, shimmer, glitter, chrome, holographic]
        - in: query
          name: inStock
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of nail colors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NailColorListResponse"

    post:
      operationId: createNailColor
      tags: [nail-colors]
      summary: Add nail color to inventory
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNailColorRequest"
      responses:
        "201":
          description: Nail color added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NailColor"
  /salons/{salonId}/nail-colors/{colorId}:
    get:
      operationId: getNailColor
      tags: [nail-colors]
      summary: Get nail color details
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/colorId"
      responses:
        "200":
          description: Nail color details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NailColor"

    put:
      operationId: updateNailColor
      tags: [nail-colors]
      summary: Update nail color
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/colorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNailColorRequest"
      responses:
        "200":
          description: Nail color updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NailColor"

    delete:
      operationId: deleteNailColor
      tags: [nail-colors]
      summary: Remove nail color from inventory
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/colorId"
      responses:
        "200":
          description: Nail color deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /salons/{salonId}/nail-colors/try-on:
    post:
      operationId: virtualTryOn
      tags: [nail-colors]
      summary: Virtual nail try-on (generate preview)
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VirtualTryOnRequest"
      responses:
        "200":
          description: Try-on preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualTryOnResponse"

  # Loyalty Program
...
  /salons/{salonId}/notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send SMS notification
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationResponse"
  /salons/{salonId}/notifications/templates:
    get:
      operationId: listNotificationTemplates
      tags: [notifications]
      summary: List notification templates
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTemplateListResponse"

    post:
      operationId: createNotificationTemplate
      tags: [notifications]
      summary: Create notification template
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationTemplateRequest"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTemplate"

  # Analytics
...
components:
  parameters:
    salonId:
      in: path
      name: salonId
      required: true
      schema:
        type: integer
...
    colorId:
      in: path
      name: colorId
      required: true
      schema:
        type: integer
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # Auth
...
    NailColor:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        brand:
          type: string
        colorCode:
          type: string
          description: Hex color code
        finish:
          type: string
          enum: [glossy, matte, shimmer, glitter, chrome, holographic]
        collection:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        inStock:
          type: boolean
        stockQuantity:
          type: integer
          nullable: true
        price:
          type: number
          nullable: true
        isPopular:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - name
        - brand
        - colorCode
        - finish
        - inStock
        - isPopular
        - tags
        - createdAt
        - updatedAt
    CreateNailColorRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        colorCode:
          type: string
        finish:
          type: string
          enum: [glossy, matte, shimmer, glitter, chrome, holographic]
        collection:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        price:
          type: number
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - brand
        - colorCode
        - finish
    UpdateNailColorRequest:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        colorCode:
          type: string
        finish:
          type: string
          enum: [glossy, matte, shimmer, glitter, chrome, holographic]
        collection:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        price:
          type: number
        isPopular:
          type: boolean
        tags:
          type: array
          items:
            type: string
    NailColorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NailColor"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    VirtualTryOnRequest:
      type: object
      properties:
        colorId:
          type: integer
        handImageUrl:
          type: string
          description: URL of user's hand photo
      required:
        - colorId
    VirtualTryOnResponse:
      type: object
      properties:
        previewImageUrl:
          type: string
        colorId:
          type: integer
        colorName:
          type: string
        colorCode:
          type: string
      required:
        - previewImageUrl
        - colorId
        - colorName
        - colorCode

    # Loyalty
...
    SendNotificationRequest:
      type: object
      properties:
        recipientPhone:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [appointment_reminder, waitlist_update, loyalty_update, custom]
        appointmentId:
          type: integer
        waitlistEntryId:
          type: integer
      required:
        - recipientPhone
        - message
        - type
    NotificationResponse:
      type: object
      properties:
        success:
          type: boolean
        messageId:
          type: string
          nullable: true
        message:
          type: string
      required:
        - success
        - message
    NotificationTemplate:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [appointment_reminder, waitlist_update, loyalty_update, custom]
        template:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - salonId
        - name
        - type
        - template
        - isActive
        - createdAt
    CreateNotificationTemplateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [appointment_reminder, waitlist_update, loyalty_update, custom]
        template:
          type: string
      required:
        - name
        - type
        - template
    NotificationTemplateListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NotificationTemplate"
      required:
        - data

    # Analytics
...